Allow passing event handlers via msDropDown constructor
This allows operation with jQuery 1.8, which does not give access to
registered event handlers via `$.fn.data('events')` anymore.

Instead of passing event handlers as constructor parameters, we could
use `$._data(element, 'events')`, but because that belongs to jQuery's
internal API, it is not guaranteed to be supported in the future.

Explicit callback usage:

    $(selector).msDropDown({ events: { change: changeHandler } })

Implicit callback usage:

    $(selector).on('change, changeHandler)
    $(selector).msDropDown({ events: { change: true } })